第五次提交（大概完成shiro）

+++ src/com/caomingyu/blog/service/PermissionService.java
@@ -5,7 +5,7 @@ import java.util.Set;
  
 import com.caomingyu.blog.pojo.Permission;
 import com.caomingyu.blog.pojo.Role;
- 
+
 public interface PermissionService {
     public Set<String> listPermissions(String userName);
  
@@ -13,9 +13,9 @@ public interface PermissionService {
  
     public void add(Permission permission);
  
-    public void delete(Long id);
+    public void delete(int id);
  
-    public Permission get(Long id);
+    public Permission get(int id);
  
     public void update(Permission permission);
